+++ lib/rules_mcell.dart
@@ -1,0 +1,7 @@
+/// MCell rule parsers.
+/// The original Original Mcell application was written by Mirek Wojtowicz
+/// http://psoup.math.wisc.edu/mcell/rullex_gene.html
+library cellular_automata.rules.mcell;
+
+export 'package:cellular_automata/src/rules/_ca_rules.dart';
+export 'package:cellular_automata/src/rules/mcell_generations.dart';